@@ -1,33 +1,25 @@
-﻿using System;
-using ExceptionAll.Dtos;
-using ExceptionAll.Helpers;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-using System.Collections.Generic;
+﻿namespace ExceptionAll.Details;
 
-namespace ExceptionAll.Details
+public class InternalServerErrorDetails : BaseDetails
 {
-    public class InternalServerErrorDetails : ProblemDetails
+    public InternalServerErrorDetails(ActionContext context, string title = null, string message = null, List<ErrorDetail> errors = null) :
+        base(
+        context ?? throw new ArgumentNullException(nameof(context)),
+        string.IsNullOrEmpty(title) ? "Internal Server Error" : title,
+        context.HttpContext.Request.Path,
+        StatusCodes.Status500InternalServerError,
+        string.IsNullOrEmpty(message) ? "See errors or logs for more details" : message,
+        errors)
     {
-        public InternalServerErrorDetails(ActionContext context, string title = null, string message = null, List<ErrorDetail> errors = null)
-        {
-            if (context is null) throw new ArgumentNullException(nameof(context));
-            Title = string.IsNullOrEmpty(title) == false ? title :  "Internal Server Error";
-            Instance = context.HttpContext.Request.Path;
-            Status = StatusCodes.Status500InternalServerError;
-            Detail = string.IsNullOrEmpty(message) == false ? message : "See errors or logs for more details";
-            this.AddDefaultExtensionsFromContext(context, errors);
-        }
-
-        public InternalServerErrorDetails(ExceptionContext context, string title = null, string message = null, List<ErrorDetail> errors = null)
-        {
-            if (context is null) throw new ArgumentNullException(nameof(context));
-            Title = string.IsNullOrEmpty(title) == false ? title :  "Internal Server Error";
-            Instance = context.HttpContext.Request.Path;
-            Status = StatusCodes.Status500InternalServerError;
-            Detail = string.IsNullOrEmpty(message) == false ? message : "See errors or logs for more details";
-            this.AddDefaultExtensionsFromContext(context, errors);
-        }
+    }
+    public InternalServerErrorDetails(ExceptionContext context, string title = null, string message = null, List<ErrorDetail> errors = null) :
+        base(
+        context ?? throw new ArgumentNullException(nameof(context)),
+        string.IsNullOrEmpty(title) ? "Internal Server Error" : title,
+        context.HttpContext.Request.Path,
+        StatusCodes.Status500InternalServerError,
+        string.IsNullOrEmpty(message) ? "See errors or logs for more details" : message,
+        errors)
+    {
     }
 }
