@@ -1,29 +1,23 @@
-﻿using ExceptionAll.Interfaces;
-using FluentValidation;
-using Microsoft.AspNetCore.Mvc;
-using System;
+﻿namespace ExceptionAll.Validation;
 
-namespace ExceptionAll.Validation
+public class ErrorResponseValidator : AbstractValidator<IErrorResponse>
 {
-    public class ErrorResponseValidator : AbstractValidator<IErrorResponse>
+    public ErrorResponseValidator()
     {
-        public ErrorResponseValidator()
-        {
-            RuleFor(x => x.ExceptionType)
-                .Must(x => x.IsSubclassOf(typeof(Exception)))
-                .When(x => x.ExceptionType != typeof(Exception));
+        RuleFor(x => x.ExceptionType)
+            .Must(x => x.IsSubclassOf(typeof(Exception)))
+            .When(x => x.ExceptionType != typeof(Exception));
 
-            RuleFor(x => x.ExceptionType)
-                .NotNull();
+        RuleFor(x => x.ExceptionType)
+            .NotNull();
 
-            RuleFor(x => x.DetailsType)
-                .Must(x => x.IsSubclassOf(typeof(ProblemDetails)));
+        RuleFor(x => x.DetailsType)
+            .Must(x => x.IsSubclassOf(typeof(BaseDetails)));
 
-            RuleFor(x => x.DetailsType)
-                .NotNull();
+        RuleFor(x => x.DetailsType)
+            .NotNull();
 
-            RuleFor(x => x.ErrorTitle)
-                .NotEmpty();
-        }
+        RuleFor(x => x.ErrorTitle)
+            .NotEmpty();
     }
 }
